refactor: migrate request runner to fsd
